+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
-**./src-nodejs/text-colours.js**
-* Merged into main file.
+**./src-nodejs/console-output.js**
+* Moved the 'displayAsciiArt' function into the main file.
 * File header remains for now.
+
+---
+
+**./javascript.js**
+* Added header comment to 'displayAsciiArt'
+	* Same as the header in the original file.
+* Removed requirements:
+	* ./src-nodejs/text-colours
+	* ./src-nodejs/console-output